chore: remove git LFS because it's expensive for open source projects
It's weird that hosting an open source git project on GitHub is free, but git LFS for such projects costs money.
GitHub charges for every `git clone` that downloads the LFS objects, even when cloned by people that are not even memebers of the project.

This repo had 3MB of git LFS objects, and the free monthly quota of 1GB bandwidth of LFS objects was reached very quickly due to people cloning this repo.
GitHub should do something about this since it's stupid to make people use regular git for non-textual files for open source projects just because it's free.
It doesn't make sense to make regular git free and charge for the bandwidth of LFS objects that are so small.